// src/screens/RegisterScreen/types.ts
export interface IFormData {
    name: string;
    planet: string;
    country?: string | undefined;
    state?: string | undefined;
    city?: string | undefined;
    zipCode?: string | undefined;
    street?: string | undefined;
    houseNumber?: string | undefined;
    neighborhood?: string | undefined;
    otherInfo?: string | undefined;
    factoryName?: string | undefined;
    lotNumber?: string | undefined;
}